chore(selectors): #1783 Move selectors to common/
