+++ common/selectors/getHighlightContextFromSite.js
