+++ content-src/components/Spotlight/Spotlight.js
@@ -1,7 +1,7 @@
 const React = require("react");
 const {connect} = require("react-redux");
-const {justDispatch} = require("selectors/selectors");
-const getHighlightContextFromSite = require("selectors/getHighlightContextFromSite");
+const {justDispatch} = require("common/selectors/selectors");
+const getHighlightContextFromSite = require("common/selectors/getHighlightContextFromSite");
 const {actions} = require("common/action-manager");
 const SiteIcon = require("components/SiteIcon/SiteIcon");
 const LinkMenu = require("components/LinkMenu/LinkMenu");
