+++ content-test/common/selectors/selectors.test.js
@@ -2,7 +2,7 @@ const {connect} = require("react-redux");
 const React = require("react");
 const TestUtils = require("react-addons-test-utils");
 const firstRunData = require("lib/first-run-data");
-const {justDispatch, selectNewTabSites} = require("selectors/selectors");
+const {justDispatch, selectNewTabSites} = require("common/selectors/selectors");
 const {TOP_SITES_LENGTH, HIGHLIGHTS_LENGTH} = require("common/constants.js");
 const {rawMockData, createMockProvider} = require("test/test-utils");
 const fakeState = rawMockData;
