+++ content-test/test-utils.js
@@ -2,7 +2,7 @@ const React = require("react");
 const ReactDOM = require("react-dom");
 const {Provider} = require("react-redux");
 const mockData = require("lib/fake-data");
-const {selectNewTabSites} = require("selectors/selectors");
+const {selectNewTabSites} = require("common/selectors/selectors");
 const TestUtils = require("react-addons-test-utils");
 
 const DEFAULT_STORE = {
